+++ source/IncomingMessages.IntegrationTests/IncomingMessages/GivenIncomingMessagesIsReceivedWithDelegationTests.cs
@@ -16,8 +16,8 @@ using System.Diagnostics.CodeAnalysis;
 using Azure.Messaging.ServiceBus;
 using Energinet.DataHub.EDI.BuildingBlocks.Domain.Authentication;
 using Energinet.DataHub.EDI.BuildingBlocks.Domain.Models;
-using Energinet.DataHub.EDI.IncomingMessages.IntegrationTests.EventBuilders;
 using Energinet.DataHub.EDI.IncomingMessages.IntegrationTests.Fixtures;
+using Energinet.DataHub.EDI.IncomingMessages.IntegrationTests.RequestBuilders;
 using Energinet.DataHub.EDI.IncomingMessages.Interfaces;
 using Energinet.DataHub.EDI.IntegrationTests.TestDoubles;
 using Energinet.DataHub.EDI.MasterData.Interfaces;
