+++ source/IncomingMessages.IntegrationTests/RequestBuilders/RequestAggregatedMeasureDataRequestBuilder.cs
@@ -18,7 +18,7 @@ using Energinet.DataHub.EDI.BuildingBlocks.Domain.Models;
 using Energinet.DataHub.EDI.IncomingMessages.Interfaces.Models;
 using NodaTime;
 
-namespace Energinet.DataHub.EDI.IncomingMessages.IntegrationTests.EventBuilders;
+namespace Energinet.DataHub.EDI.IncomingMessages.IntegrationTests.RequestBuilders;
 
 [SuppressMessage("Design", "CA1062:Validate arguments of public methods", Justification = "Used only in tests")]
 public static class RequestAggregatedMeasureDataRequestBuilder
